Ingedients

+++ src/FoodItem.java
@@ -3,7 +3,7 @@ import java.util.ArrayList;
 public class FoodItem {
     public int Food_ID;
     public String name;
-    public double price;
+    public float price;
     public boolean vegetarian;
     public ArrayList<Ingredients> foodIngredients;
 
@@ -25,11 +25,11 @@ public class FoodItem {
         
     }
     void Price(ArrayList<Ingredients> foodIngredients){
-        double price=0;
+        float price=0;
         for (int i = 0; i < foodIngredients.size(); i++) {
             price=price+foodIngredients.get(i).price;
         }
-        price=1.4*price;
+        price=(float) (1.4*price);
         
     }
 }
